Membuat COmponent untuk Kartu Iuran

+++ database/migrations/2020_10_27_092356_create_keuangans_table.php
@@ -19,7 +19,7 @@ class CreateKeuangansTable extends Migration
             $table->integer('user_id')->unsigned();
             $table->string('email');
             $table->string('nama');
-            $table->integer('jumlah');
+            $table->bigInteger('jumlah');
             $table->string('kategori')->nullable();
             $table->timestamps();
         });
